+++ src/pandas_contract/__init__.py
@@ -2,6 +2,15 @@
 
 from ._decorator import argument, result
 from ._lib import from_arg
-from .mode import as_mode, raises, set_mode, silent
+from .mode import Modes, as_mode, raises, set_mode, silent
 
-__all__ = ["argument", "as_mode", "from_arg", "raises", "result", "set_mode", "silent"]
+__all__ = [
+    "Modes",
+    "argument",
+    "as_mode",
+    "from_arg",
+    "raises",
+    "result",
+    "set_mode",
+    "silent",
+]