+++ src/tests/unit_tests/conftest.py
@@ -1,0 +1,5 @@
+"""pytest setup module."""
+
+from pandas_contract import Modes, set_mode
+
+set_mode(Modes.RAISE)