+++ src/tests/unit_tests/test_decorator.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+import pandas as pd
+import pandera as pa
 import pytest
 
 from pandas_contract import argument, result
@@ -29,3 +31,18 @@ def test_unknown_arg(verify: argument | result) -> None:
         @argument(arg="x")
         def my_fn(a: int) -> None:
             pass
+
+
+def test_multiple_decorators() -> None:
+    """Test multiple decorators on the same function."""
+
+    @result(
+        same_index_as="df", schema=pa.DataFrameSchema({"a": pa.Column(int)}, checks=[])
+    )
+    @argument(arg="df", same_index_as="df2")
+    @argument(arg="df", same_size_as="df2")
+    @result(same_index_as="df")
+    def my_fn(df: pd.DataFrame, df2: pd.DataFrame) -> pd.DataFrame:
+        return df
+
+    my_fn(pd.DataFrame({"a": [1]}), pd.DataFrame({"a": [1]}))